refactor(analysis): split LinearSoftplusModel into LinearSoftplus + DoubleSoftplus
The single 19-parameter LinearSoftplusModel conflated two orthogonal user
questions ("do I want a saturating model?" and "do I want to fit beta?") via
a 3-mode runtime dispatch. Split it into two siblings with explicit user
intent:

- LinearSoftplus (4 params, no saturation ceiling) — pruning ON by default.
  Use when colonies are still in the linear-growth regime.
- DoubleSoftplus (4-or-5 params, with saturation ceiling) — binary mode
  dispatch (fitted_beta vs fixed_beta). Use when there is a real plateau.

Shared infrastructure (sigma resolution, s0 prior, aggregation broadcasting,
the lag-softplus math) lives in a new private base _LinearSoftplusBase.
_InoculumPrior was extracted to its own module.

Other API trims rolled in:
- s0_prior_factor's value-overload (>1=sigma, <=1=cv) replaced by mutually
  exclusive s0_prior_cv / s0_prior_sigma. Default-CV fallback (0.05) preserves
  legacy behavior when neither is set.
- Removed dead/redundant init kwargs: agg_func, v_upper, stderr_floor_quantile,
  saturation_threshold, saturation_buffer. v_upper and stderr_floor_quantile
  are now class-level constants (override via subclass).
- DoubleSoftplus renames the (formerly shared) saturation_threshold to
  shoulder_slope_ratio — disambiguated from pruning, which lives only on
  LinearSoftplus now.

Schema break: output column prefixes change from LinearSoftplusModel_* to
LinearSoftplus_* / DoubleSoftplus_*. LinearSoftplusModel is removed entirely
(no deprecation alias). All 14 call sites — tests, GUI defaults, docs — were
updated; old test_linear_softplus_model.py was split into test_linear_softplus.py
and test_double_softplus.py.

Verification: 2298 unit tests pass, ruff clean on every touched file, no new
mypy errors. Code review found a fragility in DoubleSoftplus._unpack_params
re-running shoulder detection — fixed by deriving mode from len(x).
