+++ src/phenotypic/analysis/__init__.py
@@ -6,14 +6,16 @@ than imaging geometry. Includes edge correction for grid layouts, log-phase grow
 modeling across time courses, and Tukey-style outlier removal for colony metrics.
 """
 
+from ._double_softplus import DoubleSoftplus
 from ._edge_correction import EdgeCorrector
-from ._linear_softplus import LinearSoftplusModel
+from ._linear_softplus import LinearSoftplus
 from ._log_growth_model import LogGrowthModel
 from ._tukey_outlier import TukeyOutlierRemover
 
 __all__ = [
+    "DoubleSoftplus",
     "EdgeCorrector",
-    "LinearSoftplusModel",
+    "LinearSoftplus",
     "LogGrowthModel",
     "TukeyOutlierRemover",
 ]
