+++ src/phenotypic/gui/FEATURES.md
@@ -173,7 +173,7 @@ under both standalone and dispatcher-mounted launches.
 | CLI `pipeline.json` persist   | `_persist_pipeline_to_output_dir`        | Aggregate finalize writes `<output>/pipeline.json`; `_load_pipeline_from_output_dir` prefers it over legacy.    | ✅ shipping | unit        | tests/unit/cli/test_cli_analysis.py::TestLoadPipelinePrefersCanonical::test_canonical_pipeline_json_wins |
 | CLI auto-emit `analysis.*`    | `_emit_analysis_outputs`                  | Aggregate + recompile finalize call `pipeline.analyze` and write `analysis.{csv,parquet}` when model is set.   | ✅ shipping | unit        | tests/unit/cli/test_cli_analysis.py::TestEmitAnalysisOutputs::test_writes_csv_and_parquet_when_model_configured |
 | Section param forms           | `_section_form` + `gui/_param_forms.param_form` | Each post / filter / model section card hosts a fully editable inline form generated from the analyzer's constructor signature; widget kinds dispatch by type (bool, num, str, Literal, list, tuple, multi-union). | ✅ shipping | unit | tests/unit/gui/test_param_forms.py::TestParamFormViaRegistry::test_param_form_renders_for_filter |
-| Multi-type union widget       | `_multi_union_widget`                    | Type-tag dropdown (None / true / false / number / string) + adaptive value input for params like `LinearSoftplusModel.s0_prior` (a union of `bool / float / int / str / None`). Coerces to the matching branch on save. | ✅ shipping | unit | tests/unit/gui/test_param_forms.py::TestParseWidgetValue::test_multi_union_tag_number |
+| Multi-type union widget       | `_multi_union_widget`                    | Type-tag dropdown (None / true / false / number / string) + adaptive value input for params like `LinearSoftplus.s0_prior` (a union of `bool / float / int / str / None`). Coerces to the matching branch on save. | ✅ shipping | unit | tests/unit/gui/test_param_forms.py::TestParseWidgetValue::test_multi_union_tag_number |
 | Per-section param edit save   | `_apply_param_edit` callback             | Pattern-matching callback fans every `param-*` widget edit into a fresh analyzer instance, swapped into `recipe.pipeline.{post,filters,model}` and persisted via `RecipeState.save`. Scoped to the `analysis-{kind}-{index}` prefix so it doesn't collide with builder edits. | ✅ shipping | unit | tests/unit/gui/test_param_forms.py::TestWidgetForParam::test_multi_union_renders_tagged_widget |
 | Column-aware analyzer params  | `phenotypic.tools_.ColumnRef` + `_param_forms._column_widget` | Filter / model params annotated with `ColumnRef` / `ColumnRefList` (`on`, `groupby`, `time_label`, `Kmax_label`) render as live dropdowns populated from `<output>/measurements.{parquet,csv}` via `MeasurementSchema.columns_for`. `ColumnRef \| None` unions render as a two-button RadioItems mode toggle ("Column" / "None"). Stale values (column missing from the on-disk schema) surface in a wrapper-level `title` tooltip and remain selectable so the user sees what the previous value was. | ✅ shipping | unit | tests/unit/gui/test_param_forms_column.py::TestColumnWidgets::test_scalar_dropdown_renders_for_columnref |
 | Analyzer/post registry discovery | `OperationRegistry._discover_analyzers` | `OperationRegistry.discover()` now also walks `phenotypic.analysis`, registering `SetAnalyzer` subclasses under category `"Filter"` and `ModelFitter` subclasses under `"Model"` so the analysis sub-app's section forms read param metadata from the same registry the builder uses. | ✅ shipping | unit | tests/unit/gui/test_param_forms.py::TestParamFormViaRegistry::test_edge_corrector_registered |
