+++ tests/unit/analysis/test_column_ref.py
@@ -83,7 +83,8 @@ class TestAnalyzerSignatureMarkers:
                 "LogGrowthModel",
                 {"on", "groupby", "time_label", "Kmax_label"},
             ),
-            ("LinearSoftplusModel", {"on", "groupby", "time_label"}),
+            ("LinearSoftplus", {"on", "groupby", "time_label"}),
+            ("DoubleSoftplus", {"on", "groupby", "time_label"}),
         ],
     )
     def test_subclass_marker_coverage(self, cls_name, expected_params):
