--- tests/unit/analysis/test_linear_softplus_model.py
@@ -1,1496 +1,0 @@
-"""Unit tests for :class:`phenotypic.analysis.LinearSoftplusModel`.
-
-Fixtures mirror the structure of ``test_log_growth_model.py``: synthetic
-two-group time courses built directly from ``model_func`` plus noise so
-ground-truth parameters are known. Assertions are tolerance-based to
-accommodate the least-squares-with-prior optimizer.
-"""
-
-from __future__ import annotations
-
-import numpy as np
-import pandas as pd
-import pytest
-
-from phenotypic.analysis import LinearSoftplusModel
-from phenotypic.tools_.measurement_info import (
-    LINEAR_SOFTPLUS_MODEL,
-    MODEL_METRICS,
-)
-
-
-# ---------------------------------------------------------------------- #
-# Synthetic fixtures
-# ---------------------------------------------------------------------- #
-def _build_group(
-    t: np.ndarray,
-    *,
-    v: float,
-    s0: float,
-    lam: float,
-    alpha: float,
-    smax: float | None,
-    beta: float = 10.0,
-    noise_sigma: float = 0.0,
-    n_replicates: int = 3,
-    strain: str = "Strain1",
-    rng: np.random.Generator | None = None,
-) -> pd.DataFrame:
-    """Build a replicated measurement DataFrame for a single strain."""
-    rng = rng or np.random.default_rng(0)
-    y_clean = LinearSoftplusModel.model_func(
-        t=t, v=v, s0=s0, lam=lam, alpha=alpha, smax=smax, beta=beta
-    )
-    rows = []
-    for rep in range(n_replicates):
-        noise = rng.normal(0.0, noise_sigma, size=len(t)) if noise_sigma > 0 else 0.0
-        y = np.asarray(y_clean, dtype=float) + noise
-        for ti, yi in zip(t, y):
-            rows.append(
-                {
-                    "Metadata_Time": float(ti),
-                    "Shape_Area": float(yi),
-                    "Metadata_Dataset": "Test",
-                    "Metadata_Strain": strain,
-                    "Metadata_Replicate": rep,
-                }
-            )
-    return pd.DataFrame(rows)
-
-
-@pytest.fixture(scope="module")
-def clean_fixture():
-    """Two-group noise-free fixture for parameter-recovery tests."""
-    t = np.linspace(0, 20, 30)
-    rng = np.random.default_rng(1)
-    g1 = _build_group(
-        t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-        noise_sigma=0.0, strain="Strain1", rng=rng,
-    )
-    g2 = _build_group(
-        t, v=3.0, s0=2.0, lam=6.0, alpha=10.0, smax=40.0,
-        noise_sigma=0.0, strain="Strain2", rng=rng,
-    )
-    return pd.concat([g1, g2], ignore_index=True)
-
-
-@pytest.fixture(scope="module")
-def noisy_fixture():
-    """Two-group noisy fixture for R² / schema tests."""
-    t = np.linspace(0, 20, 30)
-    rng = np.random.default_rng(2)
-    g1 = _build_group(
-        t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-        noise_sigma=0.5, strain="Strain1", rng=rng,
-    )
-    g2 = _build_group(
-        t, v=3.0, s0=2.0, lam=6.0, alpha=10.0, smax=40.0,
-        noise_sigma=0.5, strain="Strain2", rng=rng,
-    )
-    return pd.concat([g1, g2], ignore_index=True)
-
-
-# ---------------------------------------------------------------------- #
-# Basic behavior
-# ---------------------------------------------------------------------- #
-class TestBasics:
-    def test_initialization(self):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        assert m.on == "Shape_Area"
-        assert m.groupby == ["Metadata_Dataset", "Metadata_Strain"]
-        assert m.time_label == "Metadata_Time"
-        assert m.smax is None
-        assert m.beta is None
-        assert m.stderr_label is None
-        assert m.stderr_floor_quantile == 0.25
-        assert m.s0_prior is None
-        assert m.s0_prior_factor == 0.05
-        assert m.s0_prior_groupby is None
-        assert m.prune_saturated is True
-        assert m.saturation_threshold == 0.05
-        assert m.saturation_buffer == 2
-        assert m.v_upper == 50.0
-        assert m.loss == "huber"
-        assert not m.verbose
-
-    def test_model_func_shapes(self):
-        t = np.linspace(0, 10, 20)
-        y_open = LinearSoftplusModel.model_func(
-            t=t, v=2.0, s0=0.5, lam=2.0, alpha=10.0, smax=None
-        )
-        assert y_open.shape == t.shape
-        assert y_open[-1] > y_open[0]
-
-        y_sat = LinearSoftplusModel.model_func(
-            t=t, v=2.0, s0=0.5, lam=2.0, alpha=10.0, smax=5.0, beta=10.0
-        )
-        assert y_sat.shape == t.shape
-        # saturation ceiling should cap the growth
-        assert y_sat[-1] < 5.0 + 1e-6
-
-    def test_schema(self, noisy_fixture):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        results = m.analyze(noisy_fixture)
-        assert isinstance(results, pd.DataFrame)
-        assert not results.empty
-
-        expected = [
-            LINEAR_SOFTPLUS_MODEL.v,
-            LINEAR_SOFTPLUS_MODEL.s0,
-            LINEAR_SOFTPLUS_MODEL.lam,
-            LINEAR_SOFTPLUS_MODEL.alpha,
-            LINEAR_SOFTPLUS_MODEL.smax,
-            LINEAR_SOFTPLUS_MODEL.beta,
-            LINEAR_SOFTPLUS_MODEL.mode,
-            MODEL_METRICS.MAE,
-            MODEL_METRICS.MSE,
-            MODEL_METRICS.RMSE,
-            MODEL_METRICS.R2,
-            MODEL_METRICS.NUM_SAMPLES,
-            MODEL_METRICS.LOSS,
-            MODEL_METRICS.STATUS,
-        ]
-        for col in expected:
-            assert col in results.columns, f"missing column: {col}"
-
-    def test_r2_finite_and_bounded(self, noisy_fixture):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        results = m.analyze(noisy_fixture)
-        r2 = results[MODEL_METRICS.R2]
-        assert r2.notna().all()
-        assert np.isfinite(r2).all()
-        assert (r2 <= 1.0 + 1e-6).all()
-        # both groups should fit reasonably well on the noisy fixture
-        assert (r2 > 0.8).all()
-
-
-# ---------------------------------------------------------------------- #
-# Parameter recovery
-# ---------------------------------------------------------------------- #
-class TestParameterRecovery:
-    def test_recovers_ground_truth(self, clean_fixture):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        results = m.analyze(clean_fixture).set_index("Metadata_Strain")
-
-        # Strain1: v=5, s0=1, lam=4, smax=50
-        row1 = results.loc["Strain1"]
-        assert abs(row1[LINEAR_SOFTPLUS_MODEL.v] - 5.0) < 0.3
-        assert abs(row1[LINEAR_SOFTPLUS_MODEL.s0] - 1.0) < 0.5
-        assert abs(row1[LINEAR_SOFTPLUS_MODEL.lam] - 4.0) < 0.5
-
-        # Strain2: v=3, s0=2, lam=6, smax=40
-        row2 = results.loc["Strain2"]
-        assert abs(row2[LINEAR_SOFTPLUS_MODEL.v] - 3.0) < 0.3
-        assert abs(row2[LINEAR_SOFTPLUS_MODEL.s0] - 2.0) < 0.5
-        assert abs(row2[LINEAR_SOFTPLUS_MODEL.lam] - 6.0) < 0.5
-
-
-# ---------------------------------------------------------------------- #
-# Weighting
-# ---------------------------------------------------------------------- #
-class TestWeighting:
-    def test_weighted_vs_unweighted_differ(self):
-        """Feed a fixture where early-time points are deliberately noisy.
-
-        With the `stderr_label` column flagging high SE at early times,
-        the weighted fit should pull ``s0`` closer to the true value
-        than the unweighted fit.
-        """
-        t = np.linspace(0, 20, 25)
-        rng = np.random.default_rng(10)
-        clean = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng,
-        )
-        # Corrupt the first 4 timepoints (early times, replicated) to be noisy.
-        early_times = np.unique(t)[:4]
-        mask = clean["Metadata_Time"].isin(early_times)
-        clean.loc[mask, "Shape_Area"] += rng.normal(0, 5.0, size=mask.sum())
-
-        # Provide a stderr_label: high for noisy early points, low elsewhere.
-        clean["Area_SE"] = np.where(mask, 5.0, 0.1)
-
-        m_weighted = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE",
-            prune_saturated=False,
-        )
-        # To make "unweighted" comparable, supply a uniform SE column.
-        clean["Area_SE_uniform"] = 1.0
-        m_unweighted = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE_uniform",
-            prune_saturated=False,
-        )
-
-        res_w = m_weighted.analyze(clean)
-        res_u = m_unweighted.analyze(clean)
-
-        s0_w = float(res_w[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        s0_u = float(res_u[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        # The two fits must differ non-trivially — weighting has an
-        # effect on the result. A stricter "weighted is closer to
-        # truth" comparison is fragile here: under hard saturation
-        # (smax=50, 21/25 points pinned at the ceiling) ``s0`` is
-        # only weakly identifiable from the 4 lag-phase points, and
-        # scipy's trust-region can land either fit on a near-flat
-        # gradient at bound-adjacent points depending on BLAS/scipy
-        # build. We test *influence*, not strict dominance.
-        assert abs(s0_w - s0_u) > 1e-3
-
-    def test_stderr_label_column_passed_through(self):
-        """Given a stderr_label, it must be used — not auto-SEM."""
-        t = np.linspace(0, 20, 15)
-        rng = np.random.default_rng(11)
-        df = _build_group(
-            t, v=4.0, s0=1.0, lam=4.0, alpha=10.0, smax=40.0,
-            noise_sigma=0.2, strain="Strain1", rng=rng, n_replicates=2,
-        )
-        df["Area_SE"] = 1.0
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE",
-            prune_saturated=False,
-        )
-        res = m.analyze(df)
-        # Smoke check: a fit completed with finite parameters.
-        assert np.isfinite(res[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        assert np.isfinite(res[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-
-
-# ---------------------------------------------------------------------- #
-# Stderr floor (quantile-relative)
-# ---------------------------------------------------------------------- #
-class TestStderrFloor:
-    @pytest.mark.parametrize("bad", [0.0, -0.1, 1.5, np.inf, np.nan])
-    def test_validation_rejects_invalid_quantile(self, bad):
-        with pytest.raises(ValueError, match="stderr_floor_quantile"):
-            LinearSoftplusModel(
-                on="Shape_Area",
-                groupby=["Metadata_Dataset", "Metadata_Strain"],
-                stderr_floor_quantile=bad,
-            )
-
-    @pytest.mark.parametrize("good", [0.01, 0.25, 0.5, 1.0])
-    def test_validation_accepts_valid_quantile(self, good):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_floor_quantile=good,
-        )
-        assert m.stderr_floor_quantile == good
-
-    def test_floor_neutralizes_tiny_sigma_point(self):
-        """Tiny σ on a corrupted mid-curve point should not pin the fit.
-
-        Without a floor, a single point with σ two orders of magnitude
-        below the rest dominates the 1/σ² weighting and drags the
-        parameters. With ``stderr_floor_quantile=0.25``, its weight is
-        capped and the fit recovers parameters closer to truth.
-        """
-        t = np.linspace(0, 20, 25)
-        rng = np.random.default_rng(42)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.2, strain="Strain1", rng=rng, n_replicates=3,
-        )
-
-        # Corrupt one mid-curve timepoint (during the growth transition)
-        # with a large y bump *and* give it an artificially tiny σ.
-        times = np.sort(df["Metadata_Time"].unique())
-        corrupt_t = float(times[len(times) // 2])
-        mask = df["Metadata_Time"] == corrupt_t
-        df.loc[mask, "Shape_Area"] += 15.0
-
-        df["Area_SE"] = np.where(mask, 0.001, 0.5)
-
-        common = dict(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE",
-            prune_saturated=False,
-        )
-        res_no_floor = LinearSoftplusModel(
-            **common, stderr_floor_quantile=None,
-        ).analyze(df)
-        res_floor = LinearSoftplusModel(**common).analyze(df)
-
-        v_no = float(res_no_floor[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        v_fl = float(res_floor[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        lam_no = float(res_no_floor[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-        lam_fl = float(res_floor[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-
-        # Floored fit should be closer to ground truth on at least one
-        # of the two most-affected parameters (v=5.0, lam=4.0).
-        better_v = abs(v_fl - 5.0) < abs(v_no - 5.0)
-        better_lam = abs(lam_fl - 4.0) < abs(lam_no - 4.0)
-        assert better_v or better_lam, (
-            f"Neither v nor lam improved with floor: "
-            f"v_no={v_no:.3f} v_fl={v_fl:.3f} "
-            f"lam_no={lam_no:.3f} lam_fl={lam_fl:.3f}"
-        )
-        # Fits must actually differ — the floor should have done
-        # something.
-        assert abs(v_no - v_fl) + abs(lam_no - lam_fl) > 1e-3
-
-    def test_none_explicitly_disables_floor(self):
-        """Explicit ``None`` skips the quantile floor entirely.
-
-        The default (``0.25``) and ``None`` must produce different
-        fits on a fixture where one timepoint has σ an order of
-        magnitude below the rest — otherwise the floor is a no-op and
-        the ``None`` escape hatch is meaningless.
-        """
-        t = np.linspace(0, 20, 25)
-        rng = np.random.default_rng(3)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.2, strain="Strain1", rng=rng, n_replicates=3,
-        )
-        times = np.sort(df["Metadata_Time"].unique())
-        corrupt_t = float(times[len(times) // 2])
-        mask = df["Metadata_Time"] == corrupt_t
-        df.loc[mask, "Shape_Area"] += 10.0
-        df["Area_SE"] = np.where(mask, 0.01, 0.5)
-
-        common = dict(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE",
-            prune_saturated=False,
-        )
-        res_default = LinearSoftplusModel(**common).analyze(df)
-        res_none = LinearSoftplusModel(
-            **common, stderr_floor_quantile=None,
-        ).analyze(df)
-
-        v_default = float(res_default[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        v_none = float(res_none[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        lam_default = float(res_default[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-        lam_none = float(res_none[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-        # Default floor and explicit None must differ on this fixture.
-        assert abs(v_default - v_none) + abs(lam_default - lam_none) > 1e-3
-
-    def test_floor_noop_when_sigma_uniform(self, noisy_fixture):
-        """With uniform σ, default floor lifts nothing — a true no-op.
-
-        When every σ in the group is identical, the 25th-percentile
-        floor equals the σ itself, so ``np.maximum(sigma, floor)`` is
-        a no-op. The default and ``None`` must produce byte-identical
-        fits in this case.
-        """
-        df = noisy_fixture.copy()
-        df["Area_SE"] = 0.5
-
-        common = dict(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_label="Area_SE",
-            prune_saturated=False,
-        )
-        res_default = LinearSoftplusModel(**common).analyze(df)
-        res_none = LinearSoftplusModel(
-            **common, stderr_floor_quantile=None,
-        ).analyze(df)
-
-        pd.testing.assert_frame_equal(res_default, res_none)
-
-    def test_singleton_replicate_groups_fall_back_to_unweighted(self):
-        """Singleton-replicate groups should not blow up the fit.
-
-        ``groupby.transform("sem")`` returns NaN for singletons. In
-        that situation ``_resolve_y_stderr`` skips the weighting
-        pathway entirely (no positive σ means no σ information), so
-        the fit falls back to an unweighted residual. The quantile
-        floor is inapplicable here — there are no positive σ to
-        compute a quantile from — but the fit must still succeed and
-        recover ground-truth parameters within tolerance.
-        """
-        t = np.linspace(0, 20, 25)
-        rng = np.random.default_rng(7)
-        df = _build_group(
-            t, v=4.0, s0=1.0, lam=5.0, alpha=10.0, smax=40.0,
-            noise_sigma=0.3, strain="Strain1", rng=rng, n_replicates=1,
-        )
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_floor_quantile=0.25,
-            prune_saturated=False,
-        )
-        res = m.analyze(df)
-
-        # Parameters are finite and reasonably close to truth despite
-        # the all-NaN auto-SEM column.
-        v_fit = float(res[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        s0_fit = float(res[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        lam_fit = float(res[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-        assert np.isfinite(v_fit)
-        assert np.isfinite(s0_fit)
-        assert np.isfinite(lam_fit)
-        assert abs(v_fit - 4.0) < 1.0
-        assert abs(lam_fit - 5.0) < 2.0
-
-    def test_mixed_singleton_and_multi_replicate_broadcasts_pool(self):
-        """Pool column broadcasts: finite for mixed, NaN for all-singleton.
-
-        For a fit group containing both n=1 and n≥2 timepoints, the
-        pooled point-level std (median of per-timepoint stds across
-        the n≥2 timepoints) is broadcast onto every row of the group.
-        For a fully-singleton fit group, the pool column is NaN — the
-        existing unweighted-fallback path still applies.
-        """
-        t = np.linspace(0, 20, 12)
-        rng = np.random.default_rng(77)
-
-        # StrainA: mixed — drop 2 of 3 replicates at alternating
-        # timepoints to make those timepoints n=1.
-        df_a = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.3, strain="StrainA", rng=rng, n_replicates=3,
-        )
-        times = np.sort(df_a["Metadata_Time"].unique())
-        singleton_times = times[::2]
-        keep_mask = ~(
-            df_a["Metadata_Time"].isin(singleton_times)
-            & (df_a["Metadata_Replicate"] > 0)
-        )
-        df_a = df_a[keep_mask].reset_index(drop=True)
-
-        # StrainB: fully-singleton.
-        df_b = _build_group(
-            t, v=3.0, s0=2.0, lam=6.0, alpha=10.0, smax=40.0,
-            noise_sigma=0.3, strain="StrainB", rng=rng, n_replicates=1,
-        )
-
-        df = pd.concat([df_a, df_b], ignore_index=True)
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        m.analyze(df)
-
-        # _latest_measurements exposes the post-broadcast raw data
-        # (set by ModelFitter.analyze) — the pool column lives here
-        # before the per-timepoint aggregation collapses it.
-        cached = m._latest_measurements
-        pool_col = "Shape_Area_std_pool"
-        assert pool_col in cached.columns
-
-        pool_a = cached.loc[cached["Metadata_Strain"] == "StrainA", pool_col]
-        pool_b = cached.loc[cached["Metadata_Strain"] == "StrainB", pool_col]
-        # StrainA has n=3 timepoints → finite positive pool broadcast
-        # to every row of the fit group.
-        assert pool_a.notna().all()
-        assert (pool_a > 0).all()
-        assert pool_a.nunique() == 1  # broadcast constant within group
-        # StrainB has only n=1 timepoints → pool cannot be estimated.
-        assert pool_b.isna().all()
-
-    def test_mixed_singleton_pool_beats_epsilon_baseline_on_s0(self):
-        """Pool fallback recovers ``s0`` better than the old ε-fill.
-
-        Build a mixed n=1/n=3 fit group and compare two fits, both
-        with ``stderr_floor_quantile=None`` to isolate the fill
-        behavior:
-
-        - **Baseline** — supply ``stderr_label`` with σ ≈ 1e-9 on n=1
-          rows and a SEM-sized σ on n=3 rows. This mimics the pre-fix
-          auto-path where n=1 σ collapsed to ε and the 1/σ² weight
-          made those rows dominate the fit (interpolating their
-          noise). Because ``s0`` is only identifiable from the lag-
-          phase timepoints, it's the parameter most sensitive to this
-          weighting pathology.
-        - **Pool** — use the auto-path (``stderr_label=None``), so
-          n=1 rows get σ ≈ pooled point-level std. All rows contribute
-          comparably to the fit.
-
-        Expectation: the pool-path ``s0`` lands at least as close to
-        truth as the baseline ``s0``. This is a genuine regression
-        guard — the old code, with its ε fill, cannot produce the
-        pool-path behavior.
-        """
-        t = np.linspace(0, 20, 20)
-        rng = np.random.default_rng(321)
-        df_full = _build_group(
-            t, v=4.0, s0=1.0, lam=4.0, alpha=10.0, smax=40.0,
-            noise_sigma=0.3, strain="Strain1", rng=rng, n_replicates=3,
-        )
-        times = np.sort(df_full["Metadata_Time"].unique())
-        singleton_times = times[::2]
-        keep_mask = ~(
-            df_full["Metadata_Time"].isin(singleton_times)
-            & (df_full["Metadata_Replicate"] > 0)
-        )
-        df = df_full[keep_mask].reset_index(drop=True)
-
-        # Per-row replicate count at each timepoint (1 for singletons,
-        # 3 for the preserved triplet timepoints).
-        n_at_t = df.groupby("Metadata_Time")["Shape_Area"].transform("count")
-        sem_like = 0.3 / np.sqrt(3)  # true SEM scale for n=3 rows
-        eps_sigma = 1e-9  # mimics the pre-fix ε fill on n=1 rows
-        df["sigma_baseline"] = np.where(n_at_t == 1, eps_sigma, sem_like)
-
-        common = dict(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            stderr_floor_quantile=None,
-            prune_saturated=False,
-        )
-        baseline = LinearSoftplusModel(
-            **common, stderr_label="sigma_baseline",
-        ).analyze(df)
-        pool = LinearSoftplusModel(**common).analyze(df)
-
-        s0_baseline = float(baseline[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        s0_pool = float(pool[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        v_pool = float(pool[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        lam_pool = float(pool[LINEAR_SOFTPLUS_MODEL.lam].iloc[0])
-
-        # Both fits produce finite parameters.
-        assert np.isfinite(s0_baseline)
-        assert np.isfinite(s0_pool)
-
-        # Pool-path s0 is at least as close to truth as the baseline.
-        # Small tolerance (1e-9) absorbs optimizer-level jitter when the
-        # two paths happen to converge to the same local minimum on a
-        # particularly benign fixture.
-        assert abs(s0_pool - 1.0) <= abs(s0_baseline - 1.0) + 1e-9
-
-        # Pool-path recovers v and lam within tightened bounds (< 15%
-        # and < 25% of truth respectively) — discriminating against a
-        # fit that drifted badly due to NaN-handling pathologies.
-        assert abs(v_pool - 4.0) < 0.6
-        assert abs(lam_pool - 4.0) < 1.0
-
-
-# ---------------------------------------------------------------------- #
-# smax fallback
-# ---------------------------------------------------------------------- #
-class TestSmaxFallback:
-    def test_per_group_observed_max(self, clean_fixture):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        results = m.analyze(clean_fixture).set_index("Metadata_Strain")
-        # Strain1 ground truth smax=50 → observed max ≈ 50
-        assert abs(results.loc["Strain1", LINEAR_SOFTPLUS_MODEL.smax] - 50.0) < 1.0
-        # Strain2 ground truth smax=40 → observed max ≈ 40
-        assert abs(results.loc["Strain2", LINEAR_SOFTPLUS_MODEL.smax] - 40.0) < 1.0
-
-    def test_explicit_smax_overrides(self, clean_fixture):
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            smax=100.0,
-            prune_saturated=False,
-        )
-        results = m.analyze(clean_fixture)
-        assert (results[LINEAR_SOFTPLUS_MODEL.smax] == 100.0).all()
-
-
-# ---------------------------------------------------------------------- #
-# Saturation pruning
-# ---------------------------------------------------------------------- #
-class TestSaturationPruning:
-    def test_basic_pruning_drops_plateau_and_preserves_growth(self):
-        """On a 20-growth + 20-plateau fixture, ``_prepare_group`` must
-        drop real rows (catches a silent no-op) and keep the growth
-        phase intact, and the end-to-end fit must still recover ``v``.
-
-        This combines three checks that together pin down the pruning
-        behavior deterministically:
-          1. ``len(pruned) < len(group)`` — a no-op `_prepare_group`
-             (returning the group unchanged) fails here.
-          2. ``len(pruned) >= len(t_growth)`` — an over-aggressive
-             implementation that trims the growth phase fails here.
-          3. The downstream fit recovers ``v`` near ground truth — a
-             broken pruner that e.g. drops the wrong rows would fail.
-        """
-        t_growth = np.linspace(0, 20, 20)
-        t_plateau = np.linspace(20.5, 50, 20)
-        t_all = np.concatenate([t_growth, t_plateau])
-        rng = np.random.default_rng(20)
-        df = _build_group(
-            t_all, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=True,
-        )
-        group = df.groupby(["Metadata_Dataset", "Metadata_Strain"]).get_group(
-            ("Test", "Strain1")
-        )
-        pruned = m._prepare_group(group)
-
-        # Direct behavioral check — pruning must actually drop rows.
-        assert len(pruned) < len(group), (
-            f"expected pruning to drop plateau rows; "
-            f"got {len(pruned)}/{len(group)}"
-        )
-        # Dynamic-range check — the pruned slice must span the full
-        # growth transition (lag near s0=1 through saturation near
-        # smax=50). An over-aggressive pruner that clipped mid-growth
-        # would fail here.
-        y_pruned = pruned[m.on].to_numpy(dtype=float)
-        assert y_pruned.min() <= 2.0, (
-            f"pruner dropped the lag phase; min y={y_pruned.min():.2f}"
-        )
-        assert y_pruned.max() >= 49.0, (
-            f"pruner dropped the saturation shoulder; max y={y_pruned.max():.2f}"
-        )
-
-        # End-to-end smoke — downstream fit recovers v within tolerance.
-        res = m.analyze(df)
-        v_fit = float(res[LINEAR_SOFTPLUS_MODEL.v].iloc[0])
-        assert abs(v_fit - 5.0) < 0.3, (
-            f"pruned fit did not recover v within tolerance: v_fit={v_fit:.4f}"
-        )
-
-    def test_lag_noise_does_not_trim_lag(self):
-        """Heavy lag-phase noise must NOT trigger early trimming."""
-        t = np.linspace(0, 20, 30)
-        rng = np.random.default_rng(21)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=6.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        # Add large noise only to lag-phase points (t <= 5).
-        lag_mask = df["Metadata_Time"] <= 5.0
-        df.loc[lag_mask, "Shape_Area"] += rng.normal(
-            0, 0.3, size=int(lag_mask.sum())
-        )
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=True,
-        )
-        pruned = m._prepare_group(
-            df.groupby(["Metadata_Dataset", "Metadata_Strain"]).get_group(
-                ("Test", "Strain1")
-            )
-        )
-        # All pre-peak (t <= 8) rows should survive.
-        lag_rows_kept = (pruned["Metadata_Time"] <= 5.0).sum()
-        lag_rows_original = int(lag_mask.sum())
-        assert lag_rows_kept == lag_rows_original
-
-    def test_transient_growth_dip_not_pruned(self):
-        """A single noisy low-derivative point mid-growth must NOT trigger."""
-        t = np.linspace(0, 20, 40)
-        rng = np.random.default_rng(22)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        # Inject a transient dip at a mid-growth timepoint.
-        mid_time = 8.0
-        dip_mask = df["Metadata_Time"] == df["Metadata_Time"].iloc[
-            np.argmin(np.abs(df["Metadata_Time"] - mid_time))
-        ]
-        df.loc[dip_mask, "Shape_Area"] *= 0.95  # small dip
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=True,
-        )
-        group = df.groupby(["Metadata_Dataset", "Metadata_Strain"]).get_group(
-            ("Test", "Strain1")
-        )
-        pruned = m._prepare_group(group)
-        # Ensure the dip location is still present.
-        assert pruned["Metadata_Time"].max() >= mid_time + 1.0
-
-    def test_non_saturating_curve_is_noop(self):
-        """Curve ending mid-growth must not be pruned."""
-        # Stop well before the curve saturates: t ∈ [0, 6] with smax=50.
-        t = np.linspace(0, 6, 15)
-        rng = np.random.default_rng(23)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=True,
-        )
-        group = df.groupby(["Metadata_Dataset", "Metadata_Strain"]).get_group(
-            ("Test", "Strain1")
-        )
-        pruned = m._prepare_group(group)
-        # Tail-slope guard must short-circuit: nothing pruned.
-        assert len(pruned) == len(group)
-
-
-# ---------------------------------------------------------------------- #
-# Inoculum prior
-# ---------------------------------------------------------------------- #
-class TestInoculumPrior:
-    def test_column_prior_pulls_s0_toward_group_mean(self):
-        """A tight inoc_size column far from the data's implied s0 must
-        bias the fit toward the column's group mean."""
-        t = np.linspace(0, 20, 25)
-        rng = np.random.default_rng(30)
-        df = _build_group(
-            t, v=5.0, s0=3.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.3, strain="Strain1", rng=rng, n_replicates=2,
-        )
-        # Inoculum measurements clustered tightly around 0.5 —
-        # far from the curve's implied s0 ≈ 3.
-        df["Inoc_Size"] = rng.normal(0.5, 0.02, size=len(df))
-
-        m_no_prior = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        m_prior = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            prune_saturated=False,
-        )
-
-        s0_no = float(m_no_prior.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        s0_yes = float(m_prior.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-
-        group_mean = float(df["Inoc_Size"].mean())
-        # The prior should pull s0 closer to the column's group mean
-        # than the no-prior fit.
-        assert abs(s0_yes - group_mean) < abs(s0_no - group_mean)
-
-    def test_no_inoc_label_omits_prior_residual(self):
-        """Verify residual length: N without prior kwargs, N+1 with them.
-
-        Locks in the symmetry of ``_loss_func`` — a regression in either
-        direction (always appending, or never appending) would flip one
-        of the two assertions below.
-        """
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        y = np.asarray([1.0, 2.0, 3.0], dtype=float)
-        t_arr = np.asarray([0.0, 1.0, 2.0], dtype=float)
-        params = [1.0, 0.5, 1.0, 10.0]
-
-        # Negative: no prior kwargs → residual vector matches y shape.
-        res = m._loss_func(params, t_arr, y, smax=5.0)
-        assert res.shape == (3,)
-
-        # Positive: prior kwargs appended → residual vector is length N+1.
-        res_with_prior = m._loss_func(
-            params, t_arr, y, smax=5.0, s0_prior_mean=0.5, s0_prior_sigma=0.1
-        )
-        assert res_with_prior.shape == (4,)
-
-    def test_zero_variance_column_engages_cv_floor(self):
-        """A zero-variance inoculum column must now engage the prior at
-        ``σ = cv × µ`` rather than silently dropping it. This is the
-        core behavior flip of the sigma-floor rework — the old code
-        returned ``None`` because the sample std was 0, quietly
-        disabling regularization in the most common user case
-        (``df["Inoc"] = 80`` scalar broadcast)."""
-        t = np.linspace(0, 20, 15)
-        rng = np.random.default_rng(32)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        df["Inoc_Size"] = 0.5  # zero-variance column
-        # ``beta=10.0`` pins both fits to the 4-parameter fixed-beta
-        # variant. In fitted-beta mode (5 params) on noise-free data
-        # the extra degree of freedom lets the optimizer reach the
-        # data-implied ``s0`` at numerical precision, swamping the
-        # prior's pull and making the end-to-end assertion brittle.
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            s0_prior_factor=0.2,
-            beta=10.0,
-            prune_saturated=False,
-        )
-
-        # Direct check: build the group the way analyze() would (via
-        # the prior's own prepare step) and assert the resolver now
-        # returns (µ, cv × µ) instead of None.
-        probe = df.copy()
-        m._prior.prepare(probe, m.groupby, m.time_label)
-        group = probe.groupby(
-            ["Metadata_Dataset", "Metadata_Strain"]
-        ).get_group(("Test", "Strain1"))
-        stats = m._inoc_stats(group)
-        assert stats is not None
-        mu, sigma = stats
-        assert mu == pytest.approx(0.5)
-        assert sigma == pytest.approx(0.2 * 0.5)
-
-        # End-to-end: compare against a no-prior fit on the same data.
-        # With the prior engaged the fitted s0 must land *strictly*
-        # closer to µ=0.5 than an unregularized fit would.
-        m_no_prior = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            beta=10.0,
-            prune_saturated=False,
-        )
-        s0_no = float(
-            m_no_prior.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0]
-        )
-        s0_yes = float(m.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0])
-        assert np.isfinite(s0_yes)
-        assert abs(s0_yes - 0.5) < abs(s0_no - 0.5)
-
-    def test_direct_mean_scalar_prior(self):
-        """A positive scalar ``s0_prior`` must engage the prior
-        uniformly across groups without any column in the data."""
-        t = np.linspace(0, 20, 20)
-        rng = np.random.default_rng(40)
-        df = _build_group(
-            t, v=5.0, s0=3.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.2, strain="Strain1", rng=rng, n_replicates=2,
-        )
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior=0.5,
-            s0_prior_factor=0.1,
-            prune_saturated=False,
-        )
-        # stats_for should return (0.5, 0.05) regardless of group content
-        group = df.groupby(
-            ["Metadata_Dataset", "Metadata_Strain"]
-        ).get_group(("Test", "Strain1"))
-        stats = m._inoc_stats(group)
-        assert stats == pytest.approx((0.5, 0.05))
-
-        # End-to-end: the prior pulls s0 toward 0.5, away from the
-        # data-implied s0=3.0.
-        m_no_prior = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        s0_no = float(
-            m_no_prior.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0]
-        )
-        s0_yes = float(
-            m.analyze(df)[LINEAR_SOFTPLUS_MODEL.s0].iloc[0]
-        )
-        assert abs(s0_yes - 0.5) < abs(s0_no - 0.5)
-
-    def test_direct_sigma_override(self):
-        """``s0_prior_factor > 1`` selects the absolute-σ branch,
-        overriding the CV-derived σ."""
-        t = np.linspace(0, 20, 15)
-        rng = np.random.default_rng(41)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        df["Inoc_Size"] = 10.0
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            s0_prior_factor=2.5,       # > 1 → absolute σ
-            prune_saturated=False,
-        )
-        probe = df.copy()
-        m._prior.prepare(probe, m.groupby, m.time_label)
-        group = probe.groupby(
-            ["Metadata_Dataset", "Metadata_Strain"]
-        ).get_group(("Test", "Strain1"))
-        stats = m._inoc_stats(group)
-        assert stats is not None
-        mu, sigma = stats
-        assert mu == pytest.approx(10.0)
-        assert sigma == pytest.approx(2.5)
-
-    def test_s0_prior_groupby_pools_across_fit_groups(self):
-        """``s0_prior_groupby`` coarser than ``groupby`` pools µ across
-        strains within a media so every fit group sees the same µ."""
-        t = np.linspace(0, 20, 15)
-        rng = np.random.default_rng(42)
-        g1 = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.1, strain="StrainA", rng=rng, n_replicates=2,
-        )
-        g2 = _build_group(
-            t, v=4.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.1, strain="StrainB", rng=rng, n_replicates=2,
-        )
-        df = pd.concat([g1, g2], ignore_index=True)
-        # Strain-level inoc at t=0: StrainA ≈ 0.3, StrainB ≈ 0.9.
-        # Pooling across strains within the dataset should give
-        # median ≈ 0.6.
-        rng_inoc = np.random.default_rng(43)
-        df["Inoc_Size"] = np.where(
-            df["Metadata_Strain"] == "StrainA",
-            rng_inoc.normal(0.3, 0.01, size=len(df)),
-            rng_inoc.normal(0.9, 0.01, size=len(df)),
-        )
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            s0_prior_groupby=["Metadata_Dataset"],
-            prune_saturated=False,
-        )
-        probe = df.copy()
-        m._prior.prepare(probe, m.groupby, m.time_label)
-
-        t_min = probe["Metadata_Time"].min()
-        expected_mu = float(
-            probe.loc[probe["Metadata_Time"] == t_min, "Inoc_Size"].median()
-        )
-
-        for strain in ("StrainA", "StrainB"):
-            group = probe.groupby(
-                ["Metadata_Dataset", "Metadata_Strain"]
-            ).get_group(("Test", strain))
-            stats = m._inoc_stats(group)
-            assert stats is not None
-            mu, _ = stats
-            assert mu == pytest.approx(expected_mu, rel=1e-6)
-
-        # Contrast with the un-pooled behavior: if we had used the fit
-        # groupby for the prior, StrainA and StrainB would get
-        # different µ's, both far from the pooled median. This
-        # assertion proves the pooling actually changed the answer.
-        strain_local_mus = {
-            strain: float(
-                df.loc[
-                    (df["Metadata_Strain"] == strain)
-                    & (df["Metadata_Time"] == t_min),
-                    "Inoc_Size",
-                ].median()
-            )
-            for strain in ("StrainA", "StrainB")
-        }
-        assert abs(strain_local_mus["StrainA"] - expected_mu) > 0.1
-        assert abs(strain_local_mus["StrainB"] - expected_mu) > 0.1
-
-    def test_s0_prior_groupby_must_be_subset_of_groupby(self):
-        """``s0_prior_groupby`` must be a subset of ``groupby`` —
-        otherwise the empirical-Bayes hierarchy is undefined."""
-        t = np.linspace(0, 20, 10)
-        rng = np.random.default_rng(44)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        df["Inoc_Size"] = 0.5
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            s0_prior_groupby=["Metadata_Dataset"],  # not in groupby
-            prune_saturated=False,
-        )
-        with pytest.raises(ValueError, match="must be a subset"):
-            m.analyze(df)
-
-    def test_s0_prior_groupby_without_column_raises(self):
-        """``s0_prior_groupby`` requires a column-backed prior —
-        scalar or disabled priors have no column to aggregate over."""
-        with pytest.raises(ValueError, match="requires a column-backed"):
-            LinearSoftplusModel(
-                on="Shape_Area",
-                groupby=["Metadata_Dataset", "Metadata_Strain"],
-                s0_prior=0.5,
-                s0_prior_groupby=["Metadata_Dataset"],
-            )
-
-    def test_missing_s0_prior_column_raises(self):
-        """``analyze`` must fail loud when ``s0_prior`` names a column
-        absent from the data, rather than silently dropping the prior."""
-        t = np.linspace(0, 20, 10)
-        rng = np.random.default_rng(45)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.0, strain="Strain1", rng=rng, n_replicates=1,
-        )
-        # Note: df has no "Inoc_Size" column.
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior="Inoc_Size",
-            prune_saturated=False,
-        )
-        with pytest.raises(ValueError, match="not present in data columns"):
-            m.analyze(df)
-
-    def test_non_positive_factor_raises(self):
-        """``s0_prior_factor`` must be strictly positive — zero or
-        negative values would silently disengage the prior via the
-        downstream ``σ ≤ 0`` gate and reintroduce the original bug."""
-        for bad_factor in (0.0, -1.0, -0.5):
-            with pytest.raises(
-                ValueError, match="s0_prior_factor must be a positive"
-            ):
-                LinearSoftplusModel(
-                    on="Shape_Area",
-                    groupby=["Metadata_Dataset", "Metadata_Strain"],
-                    s0_prior="Inoc_Size",
-                    s0_prior_factor=bad_factor,
-                )
-
-    def test_non_positive_scalar_prior_raises(self):
-        """A scalar ``s0_prior`` must be positive — inoculum sizes are
-        physically positive, and µ ≤ 0 would silently disengage the
-        prior."""
-        for bad_mean in (0.0, -1.0):
-            with pytest.raises(
-                ValueError, match="s0_prior scalar must be a positive"
-            ):
-                LinearSoftplusModel(
-                    on="Shape_Area",
-                    groupby=["Metadata_Dataset", "Metadata_Strain"],
-                    s0_prior=bad_mean,
-                )
-
-    def test_empty_s0_prior_groupby_raises(self):
-        """``s0_prior_groupby=[]`` is rejected — pass ``None`` to fall
-        back to the fit groupby."""
-        with pytest.raises(ValueError, match="must not be an empty list"):
-            LinearSoftplusModel(
-                on="Shape_Area",
-                groupby=["Metadata_Dataset", "Metadata_Strain"],
-                s0_prior="Inoc_Size",
-                s0_prior_groupby=[],
-            )
-
-    # ------------------------------------------------------------------ #
-    # New polymorphic-dispatch behavior
-    # ------------------------------------------------------------------ #
-    def test_s0_prior_true_grounds_on_on_column(self):
-        """``s0_prior=True`` grounds µ on ``self.on`` at ``t_min``.
-
-        Under the hood, the helper sets ``label = on_column``, so
-        the prior takes its µ from the median of the fit's own
-        target column at the earliest observed timepoint.
-        """
-        t = np.linspace(0, 20, 20)
-        rng = np.random.default_rng(50)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0,
-            noise_sigma=0.1, strain="Strain1", rng=rng, n_replicates=2,
-        )
-
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior=True,
-            prune_saturated=False,
-        )
-        assert m._prior.label == "Shape_Area"
-        assert m._prior.direct_mean is None
-
-        probe = df.copy()
-        m._prior.prepare(probe, m.groupby, m.time_label)
-        group = probe.groupby(
-            ["Metadata_Dataset", "Metadata_Strain"]
-        ).get_group(("Test", "Strain1"))
-        stats = m._inoc_stats(group)
-        assert stats is not None
-        mu, sigma = stats
-
-        t_min = float(df["Metadata_Time"].min())
-        expected_mu = float(
-            df.loc[df["Metadata_Time"] == t_min, "Shape_Area"].median()
-        )
-        assert mu == pytest.approx(expected_mu, rel=1e-6)
-        assert sigma == pytest.approx(0.05 * expected_mu)  # default CV
-
-    def test_s0_prior_true_not_interpreted_as_numeric(self):
-        """Guard against the ``isinstance(True, int) is True`` gotcha:
-        ``s0_prior=True`` must take the data-grounded branch, not the
-        scalar branch with µ=1.0."""
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior=True,
-        )
-        assert m._prior.label == "Shape_Area"
-        assert m._prior.direct_mean is None  # NOT 1.0
-
-    def test_s0_prior_false_disables_prior(self):
-        """``s0_prior=False`` is an explicit-disable, semantically
-        identical to ``None``."""
-        m_false = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior=False,
-        )
-        m_none = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            s0_prior=None,
-        )
-        assert m_false._prior.is_configured is False
-        assert m_none._prior.is_configured is False
-        assert m_false._prior.label is None
-        assert m_false._prior.direct_mean is None
-
-    def test_s0_prior_invalid_type_raises(self):
-        """Non-supported types on ``s0_prior`` must raise
-        :class:`TypeError` at construction."""
-        for bad in ([1, 2], {"x": 1}, (0.5,)):
-            with pytest.raises(TypeError, match="s0_prior must be"):
-                LinearSoftplusModel(
-                    on="Shape_Area",
-                    groupby=["Metadata_Dataset", "Metadata_Strain"],
-                    s0_prior=bad,  # type: ignore[arg-type]
-                )
-
-    def test_s0_prior_factor_boundary(self):
-        """``factor=1.0`` is the CV/σ boundary: inclusive of CV, so
-        ``σ = 1.0 × µ``. Any value above (e.g. 1.0001) flips to the
-        absolute-σ branch."""
-        m_cv = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["G"],
-            s0_prior=10.0,
-            s0_prior_factor=1.0,       # ≤ 1 → CV path
-        )
-        assert m_cv._prior.cv == 1.0
-        assert m_cv._prior.direct_sigma is None
-        stats_cv = m_cv._prior.stats_for(pd.DataFrame({"G": ["x"]}))
-        assert stats_cv == pytest.approx((10.0, 10.0))  # σ = 1.0 × 10
-
-        m_sigma = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["G"],
-            s0_prior=10.0,
-            s0_prior_factor=1.0001,    # > 1 → σ path
-        )
-        assert m_sigma._prior.cv is None
-        assert m_sigma._prior.direct_sigma == pytest.approx(1.0001)
-        stats_sigma = m_sigma._prior.stats_for(pd.DataFrame({"G": ["x"]}))
-        assert stats_sigma == pytest.approx((10.0, 1.0001))
-
-    def test_s0_prior_factor_below_1_is_cv(self):
-        """Redundancy check for the CV branch: ``factor=0.1``,
-        ``µ=5`` → ``σ = 0.5``."""
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["G"],
-            s0_prior=5.0,
-            s0_prior_factor=0.1,
-        )
-        stats = m._prior.stats_for(pd.DataFrame({"G": ["x"]}))
-        assert stats == pytest.approx((5.0, 0.5))
-
-    def test_s0_prior_factor_above_1_is_sigma(self):
-        """Redundancy check for the σ branch: ``factor=3.0``, ``µ=10``
-        → ``σ = 3.0`` (independent of µ)."""
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["G"],
-            s0_prior=10.0,
-            s0_prior_factor=3.0,
-        )
-        stats = m._prior.stats_for(pd.DataFrame({"G": ["x"]}))
-        assert stats == pytest.approx((10.0, 3.0))
-
-
-# ---------------------------------------------------------------------- #
-# Degenerate input → NaN row
-# ---------------------------------------------------------------------- #
-class TestDegenerateInput:
-    def test_all_zero_data_does_not_crash(self):
-        """All-zero measurements must not raise; analyze returns a row.
-
-        The optimizer converges to some trivial parameter vector here
-        (residuals are degenerate), so we only assert that the pipeline
-        completes and produces a single row — not that the values are
-        NaN.
-        """
-        t = np.linspace(0, 20, 15)
-        rows = []
-        for rep in range(2):
-            for ti in t:
-                rows.append(
-                    {
-                        "Metadata_Time": float(ti),
-                        "Shape_Area": 0.0,
-                        "Metadata_Dataset": "Test",
-                        "Metadata_Strain": "Dead",
-                        "Metadata_Replicate": rep,
-                    }
-                )
-        df = pd.DataFrame(rows)
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        res = m.analyze(df)
-        assert len(res) == 1
-        # v must respect the configured upper bound.
-        assert 0.0 <= float(res[LINEAR_SOFTPLUS_MODEL.v].iloc[0]) <= 50.0
-
-    def test_nan_input_triggers_nan_row(self):
-        """NaN measurements propagate into the optimizer, triggering the
-        NaN fit-column path (``_nan_fit_columns``) instead of crashing."""
-        t = np.linspace(0, 20, 15)
-        rows = []
-        for ti in t:
-            rows.append(
-                {
-                    "Metadata_Time": float(ti),
-                    "Shape_Area": float("nan"),
-                    "Metadata_Dataset": "Test",
-                    "Metadata_Strain": "Broken",
-                }
-            )
-        df = pd.DataFrame(rows)
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        res = m.analyze(df)
-        assert len(res) == 1
-        # Every fitted parameter must be NaN.
-        assert np.isnan(float(res[LINEAR_SOFTPLUS_MODEL.v].iloc[0]))
-        assert np.isnan(float(res[LINEAR_SOFTPLUS_MODEL.s0].iloc[0]))
-        assert np.isnan(float(res[LINEAR_SOFTPLUS_MODEL.lam].iloc[0]))
-        assert np.isnan(float(res[LINEAR_SOFTPLUS_MODEL.alpha].iloc[0]))
-        assert pd.isna(res[LINEAR_SOFTPLUS_MODEL.beta].iloc[0])
-        # mode is a string-valued column — NaN fallback on fit failure.
-        assert pd.isna(res[LINEAR_SOFTPLUS_MODEL.mode].iloc[0])
-
-
-# ---------------------------------------------------------------------- #
-# Per-group mode dispatch
-# ---------------------------------------------------------------------- #
-class TestModeDispatch:
-    """Mode selection across the three variants — unclamped, fixed_beta,
-    fitted_beta — and how ``self.beta`` / ``self.smax`` steer the choice.
-    """
-
-    def test_saturating_curve_uses_fitted_beta(self):
-        """A curve with a clear shoulder and default ``beta=None`` picks
-        ``fitted_beta`` and recovers beta within the configured bounds."""
-        t = np.linspace(0, 20, 30)
-        rng = np.random.default_rng(100)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=12.0,
-            noise_sigma=0.1, strain="Saturated", rng=rng, n_replicates=2,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        res = m.analyze(df)
-        assert res[LINEAR_SOFTPLUS_MODEL.mode].iloc[0] == "fitted_beta"
-        beta_fit = float(res[LINEAR_SOFTPLUS_MODEL.beta].iloc[0])
-        assert 2.0 <= beta_fit <= 50.0
-        assert np.isfinite(float(res[LINEAR_SOFTPLUS_MODEL.smax].iloc[0]))
-
-    def test_non_saturating_no_smax_uses_unclamped(self):
-        """Truncated-before-saturation + ``smax=None`` + ``beta=None``
-        selects the unclamped variant and reports NaN for smax/beta."""
-        # t range cut off well before the curve would reach smax=50.
-        t = np.linspace(0, 6, 15)
-        rng = np.random.default_rng(101)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=3.0, alpha=10.0, smax=50.0, beta=10.0,
-            noise_sigma=0.0, strain="Open", rng=rng, n_replicates=1,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-        res = m.analyze(df)
-        assert res[LINEAR_SOFTPLUS_MODEL.mode].iloc[0] == "unclamped"
-        assert pd.isna(res[LINEAR_SOFTPLUS_MODEL.smax].iloc[0])
-        assert pd.isna(res[LINEAR_SOFTPLUS_MODEL.beta].iloc[0])
-
-    def test_non_saturating_with_smax_uses_fixed_beta(self):
-        """No shoulder but user-supplied ``smax`` → ``fixed_beta`` with
-        the module-default beta."""
-        t = np.linspace(0, 6, 15)
-        rng = np.random.default_rng(102)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=3.0, alpha=10.0, smax=50.0, beta=10.0,
-            noise_sigma=0.0, strain="OpenWithSmax", rng=rng, n_replicates=1,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            smax=50.0,
-        )
-        res = m.analyze(df)
-        assert res[LINEAR_SOFTPLUS_MODEL.mode].iloc[0] == "fixed_beta"
-        assert float(res[LINEAR_SOFTPLUS_MODEL.smax].iloc[0]) == 50.0
-        assert float(res[LINEAR_SOFTPLUS_MODEL.beta].iloc[0]) == 10.0
-
-    def test_explicit_beta_forces_fixed_mode_even_with_shoulder(self):
-        """User-explicit ``beta=<scalar>`` pins the mode to ``fixed_beta``
-        regardless of whether a shoulder is present."""
-        t = np.linspace(0, 20, 30)
-        rng = np.random.default_rng(103)
-        df = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=10.0,
-            noise_sigma=0.0, strain="SatButPinned", rng=rng, n_replicates=1,
-        )
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            beta=7.0,
-        )
-        res = m.analyze(df)
-        assert res[LINEAR_SOFTPLUS_MODEL.mode].iloc[0] == "fixed_beta"
-        assert float(res[LINEAR_SOFTPLUS_MODEL.beta].iloc[0]) == 7.0
-
-    def test_fitted_beta_recovers_distinct_ground_truth(self):
-        """Two saturating groups with very different true betas must
-        produce distinctly different fitted betas."""
-        t = np.linspace(0, 20, 40)
-        rng = np.random.default_rng(104)
-        g_sharp = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=25.0,
-            noise_sigma=0.05, strain="SharpKnee", rng=rng, n_replicates=2,
-        )
-        g_soft = _build_group(
-            t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=4.0,
-            noise_sigma=0.05, strain="SoftKnee", rng=rng, n_replicates=2,
-        )
-        df = pd.concat([g_sharp, g_soft], ignore_index=True)
-        m = LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-            prune_saturated=False,
-        )
-        res = m.analyze(df).set_index("Metadata_Strain")
-        for strain in ("SharpKnee", "SoftKnee"):
-            assert res.loc[strain, LINEAR_SOFTPLUS_MODEL.mode] == "fitted_beta"
-        beta_sharp = float(res.loc["SharpKnee", LINEAR_SOFTPLUS_MODEL.beta])
-        beta_soft = float(res.loc["SoftKnee", LINEAR_SOFTPLUS_MODEL.beta])
-        # The two fitted betas should be well-separated — sharp knee
-        # tends to the upper bound, soft knee toward its ground truth.
-        assert beta_sharp > beta_soft + 3.0
-        # Soft knee should land within a loose tolerance of truth.
-        assert abs(beta_soft - 4.0) < 3.0
-
-    def test_non_positive_beta_raises(self):
-        """Scalar ``beta`` must be a positive finite number."""
-        for bad in (0.0, -1.0, float("nan"), float("inf")):
-            with pytest.raises(ValueError, match="beta must be None or"):
-                LinearSoftplusModel(
-                    on="Shape_Area",
-                    groupby=["Metadata_Dataset", "Metadata_Strain"],
-                    beta=bad,
-                )
-
-
-# ---------------------------------------------------------------------- #
-# Shoulder detection
-# ---------------------------------------------------------------------- #
-class TestShoulderDetection:
-    """Unit tests for ``_has_saturation_shoulder`` — the signal driving
-    the fitted/unclamped branches of mode dispatch.
-    """
-
-    @pytest.fixture
-    def _model(self):
-        return LinearSoftplusModel(
-            on="Shape_Area",
-            groupby=["Metadata_Dataset", "Metadata_Strain"],
-        )
-
-    def _wrap(self, t, y) -> pd.DataFrame:
-        return pd.DataFrame({"Metadata_Time": t, "Shape_Area": y})
-
-    def test_saturating_curve_detected(self, _model):
-        t = np.linspace(0, 20, 30)
-        y = LinearSoftplusModel.model_func(
-            t=t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=10.0,
-        )
-        assert _model._has_saturation_shoulder(self._wrap(t, y))
-
-    def test_linear_growth_throughout_not_detected(self, _model):
-        # Pure linear growth — tail slope ≈ peak slope → no shoulder.
-        t = np.linspace(0, 20, 30)
-        y = 3.0 * t + 1.0
-        assert not _model._has_saturation_shoulder(self._wrap(t, y))
-
-    def test_flat_noise_not_detected(self, _model):
-        rng = np.random.default_rng(200)
-        t = np.linspace(0, 20, 30)
-        y = 1.0 + rng.normal(0, 1e-4, size=t.size)
-        # Dynamic-range guard should kick in.
-        assert not _model._has_saturation_shoulder(self._wrap(t, y))
-
-    def test_too_few_points_not_detected(self, _model):
-        # Even a perfect saturation shape is rejected below the
-        # minimum-samples gate.
-        t = np.linspace(0, 20, 5)
-        y = LinearSoftplusModel.model_func(
-            t=t, v=5.0, s0=1.0, lam=4.0, alpha=10.0, smax=50.0, beta=10.0,
-        )
-        assert not _model._has_saturation_shoulder(self._wrap(t, y))
-
-
-# ---------------------------------------------------------------------- #
-# LogGrowthModel regression — hooks must be no-ops by default
-# ---------------------------------------------------------------------- #
-def test_log_growth_still_works():
-    """Sanity check: adding hooks to the base class didn't break siblings."""
-    from phenotypic.analysis import LogGrowthModel
-
-    np.random.seed(0)
-    t = np.arange(0, 10, 1)
-    r_true, K_true, N0_true = 0.5, 1000.0, 50.0
-    rows = []
-    for ti in t:
-        y = K_true / (1 + (K_true - N0_true) / N0_true * np.exp(-r_true * ti))
-        for _ in range(3):
-            rows.append(
-                {
-                    "Metadata_Time": float(ti),
-                    "Shape_Area": float(y),
-                    "Metadata_Dataset": "Test",
-                    "Metadata_Strain": "Strain1",
-                }
-            )
-    df = pd.DataFrame(rows)
-    m = LogGrowthModel(
-        on="Shape_Area",
-        groupby=["Metadata_Dataset", "Metadata_Strain"],
-    )
-    res = m.analyze(df)
-    assert not res.empty
